Feat: GQL Backend allowing the query.
The commit adds the backend for queries student and school. By writing
this backend with my own hand, I am pretty much convinced about the
feature which Hasura provides. However, this is not targetted to aim
hasura but makes learning about GQL more and more clear. Currently only
`query` is allowed. Will need to tweak it further to allow the
`mutations` and connecting the stuff to any database.

Signed-off-by: Harshil Jani <harshiljani2002@gmail.com>