Feat: Added an instance of connecting PSQL DB
This commit adds the file required to just set up any form of PSQL
database quickly without much detaling about the stuff going on. Main
motive is to learn GQL backend so playing around with Databases can be
done sooner or later as a part of some other learning path.

Signed-off-by: Harshil Jani <harshiljani2002@gmail.com>